Moved terms of use default data to its own model. Deleted route for terms of use
