--- backend-admin/lib/db_list_admin_web/controllers/terms_of_use_controller.ex
@@ -1,7 +1,0 @@
-defmodule DbListAdminWeb.TermsOfUseController do
-  use DbListAdminWeb, :controller
-
-  def index(conn, _params) do
-    json conn, DbListAdmin.Resource.Database.Remapper.get_default_terms_of_use()
-  end
-end
